+++ spec/factories/pull_requests/comments.rb
@@ -4,6 +4,7 @@ FactoryBot.define do
   factory :pull_requests_comment, class: 'PullRequests::Comment' do
     external_id { '1' }
     comment_created_at { DateTime.now }
-    pull_requests_entity
+    pull_request
+    entity
   end
 end